# Conflicts:
#	frontend/src/locales/de.json
#	frontend/src/locales/en.json
#	frontend/src/locales/es.json
#	frontend/src/locales/fr.json
#	frontend/src/locales/it.json
#	frontend/src/locales/pl.json
#	frontend/src/locales/pt.json
#	frontend/src/locales/uk.json
#	frontend/src/views/LandingPage.vue
#	frontend/src/views/MemberVotePage.vue
#	frontend/src/views/PrepareSessionPage.vue
#	frontend/src/views/SessionPage.vue

+++ frontend/src/views/SessionPage.vue
@@ -1,23 +1,23 @@
 <template>
   <b-container id="session-page">
-    <b-row class="mb-3">
-      <b-col cols="8">
+    <b-row class="headers">
+      <b-col cols="auto" sm="8">
         <h1>
           {{
             planningStart
-            ? $t("page.session.during.estimation.title")
-            : $t("page.session.before.title")
+              ? $t("page.session.during.estimation.title")
+              : $t("page.session.before.title")
           }}
         </h1>
       </b-col>
-      <b-col cols="auto" class="mr-auto">
+      <b-col cols="auto">
         <copy-session-id-popup
           v-if="planningStart"
           class="float-end"
           :session-id="session_sessionID"
         />
       </b-col>
-      <b-col cols="auto">
+      <b-col id="sessionCloseCol" cols="auto">
         <session-close-button
           :is-planning-start="planningStart"
           :user-story-mode="session_userStoryMode"
@@ -26,17 +26,17 @@
     </b-row>
 
     <div v-if="!planningStart">
+      <div id="catGifDiv">
+        <b-img :src="require('@/assets/LoadingCat.gif')" class="catGif" />
+      </div>
       <copy-session-id-popup
         :text-before-session-i-d="$t('page.session.before.text.beforeID')"
         :session-id="session_sessionID"
         :text-after-session-i-d="$t('page.session.before.text.afterID')"
+        class="copy-popup"
       />
-      <h4 class="text-center m-3">
-        {{ $t("page.session.before.text.waiting") }}
-        <sub><b-icon-three-dots animation="fade" font-scale="1" /></sub>
-      </h4>
 
-      <b-row class="d-flex justify-content-center overflow-auto" style="max-height: 500px">
+      <b-row class="d-flex justify-content-center overflow-auto kick-user">
         <kick-user-wrapper
           v-for="member of members"
           :key="member.memberID"
@@ -58,13 +58,27 @@
 
     <div v-else>
       <b-row class="d-flex justify-content-start pb-3">
-        <b-col cols="auto" class="mr-auto">
-          <b-button class="mr-3" variant="outline-dark" @click="sendRestartMessage">
-            <b-icon-arrow-clockwise />
+        <b-col cols="auto" class="mr-auto optionButtonCol">
+          <b-button
+            class="mr-3 optionButton"
+            variant="outline-dark"
+            @click="
+              sendRestartMessage();
+              $event.target.blur();
+            "
+          >
+            <BIconArrowClockwise class="bIcons"></BIconArrowClockwise>
             {{ $t("page.session.during.estimation.buttons.new") }}
           </b-button>
-          <b-button class="mr-3" variant="outline-dark" @click="sendVotingFinishedMessage">
-            <b-icon-bar-chart />
+          <b-button
+            class="mr-3 optionButton"
+            variant="outline-dark"
+            @click="
+              sendVotingFinishedMessage();
+              $event.target.blur();
+            "
+          >
+            <BIconBarChartFill class="bIcons"></BIconBarChartFill>
             {{ $t("page.session.during.estimation.buttons.result") }}
           </b-button>
         </b-col>
@@ -73,7 +87,7 @@
             :start-timestamp="timerTimestamp"
             :pause-timer="estimateFinished"
             :duration="timerCountdownNumber"
-            :votingStarted="planningStart"
+            :voting-started="planningStart"
             @timerFinished="sendVotingFinishedMessage"
           />
         </b-col>
@@ -84,7 +98,7 @@
         {{ membersPending.length }} /
         {{ membersPending.length + membersEstimated.length }}
       </h4>
-      <b-row v-if="!estimateFinished" class="my-1 d-flex justify-content-center flex-wrap">
+      <b-row v-if="!estimateFinished" class="my-2 d-flex justify-content-center flex-wrap">
         <kick-user-wrapper
           v-for="member of membersPending"
           :key="member.memberID"
@@ -93,7 +107,7 @@
           :member="member"
         />
       </b-row>
-      <hr />
+      <hr class="my-5 breakingLine" />
       <h4 v-if="!session_hostVoting">
         {{ $t("page.session.during.estimation.message.finished") }}
         {{ membersEstimated.length }} /
@@ -111,7 +125,7 @@
           {{ members.length + 1 }}
         </div>
       </h4>
-      <b-row class="my-1 d-flex justify-content-center flex-wrap overflow-auto" style="max-height: 500px"
+      <b-row class="my-1 d-flex justify-content-center flex-wrap overflow-auto kick-user" style="max-height: 500px"
         v-if="highlightedMembers.includes(adminID)">
         <session-admin-card
           v-if="safedHostVoting && estimateFinished || hostEstimation !== ''"
@@ -160,15 +174,21 @@
         </div>
       </div>
     </div>
-    <b-row v-if="session_userStoryMode !== 'NO_US'" class="mt-5">
+    <b-row v-if="session_userStoryMode !== 'NO_US'" class="mt-4">
       <b-col>
         <user-story-sum-component />
       </b-col>
     </b-row>
     <b-row v-if="session_userStoryMode !== 'NO_US'">
-      <b-col cols="4">
+      <b-col v-if="!isMobile" cols="7">
         <div v-if="session_userStoryMode === 'US_JIRA'" class="refreshUserstories">
-          <b-button class="w-100 mb-3" variant="info" @click="refreshUserStories">
+          <b-button
+            class="w-100 mb-3 refreshButton"
+            @click="
+              refreshUserStories();
+              $event.target.blur();
+            "
+          >
             {{ $t("page.session.before.refreshStories") }}
           </b-button>
         </div>
@@ -182,7 +202,40 @@
           @selectedStory="onSelectedStory($event)"
         />
       </b-col>
-      <b-col cols="8">
+      <b-col v-else cols="12">
+        <div v-if="session_userStoryMode === 'US_JIRA'" class="refreshUserstories">
+          <b-button
+            class="w-100 mb-3 refreshButton"
+            @click="
+              refreshUserStories();
+              $event.target.blur();
+            "
+          >
+            {{ $t("page.session.before.refreshStories") }}
+          </b-button>
+        </div>
+        <user-stories
+          :card-set="voteSet"
+          :show-estimations="planningStart"
+          :initial-stories="userStories"
+          :show-edit-buttons="true"
+          :select-story="true"
+          @userStoriesChanged="onUserStoriesChanged"
+          @selectedStory="onSelectedStory($event)"
+        />
+      </b-col>
+    </b-row>
+    <b-row>
+      <b-col v-if="!isMobile" cols="10">
+        <user-story-descriptions
+          :card-set="voteSet"
+          :initial-stories="userStories"
+          :edit-description="true"
+          :index="index"
+          @userStoriesChanged="onUserStoriesChanged"
+        />
+      </b-col>
+      <b-col v-else cols="12">
         <user-story-descriptions
           :card-set="voteSet"
           :initial-stories="userStories"
@@ -212,6 +265,7 @@ import Project from "../model/Project";
 import KickUserWrapper from "@/components/KickUserWrapper.vue";
 import SessionCloseButton from "@/components/actions/SessionCloseButton.vue";
 import SessionStartButton from "@/components/actions/SessionStartButton.vue";
+import { BIconArrowClockwise, BIconBarChartFill } from "bootstrap-vue";
 import SessionAdminCard from "@/components/SessionAdminCard.vue";
 
 export default Vue.extend({
@@ -226,6 +280,8 @@ export default Vue.extend({
     CopySessionIdPopup,
     UserStoryDescriptions,
     NotifyHostComponent,
+    BIconArrowClockwise,
+    BIconBarChartFill,
     SessionAdminCard,
   },
   props: {
@@ -291,6 +347,11 @@ export default Vue.extend({
     },
     timerTimestamp() {
       return this.$store.state.timerTimestamp ? this.$store.state.timerTimestamp : "";
+    },
+    isMobile() {
+      return /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
+        navigator.userAgent
+      );
     },
   },
   watch: {
@@ -457,7 +518,11 @@ export default Vue.extend({
       if (this.session_userStoryMode === "US_JIRA") {
         let response;
         if (doRemove) {
-          response = await apiService.deleteUserStory(us[idx].id);
+          if (us[idx].id === null) {
+            response = 204;
+          } else {
+            response = await apiService.deleteUserStory(us[idx].id);
+          }
           us.splice(idx, 1);
           doRemove = false;
         } else {
@@ -481,6 +546,8 @@ export default Vue.extend({
         }
         if (response.status === 200) {
           this.$toast.success(this.$t("session.notification.messages.issueTrackerSynchronizeSuccess"));
+        } else if (response === 204) {
+          this.$toast.info(this.$t("session.notification.messages.issueTrackerNothingChanged"));
         } else {
           this.$toast.error(this.$t("session.notification.messages.issueTrackerSynchronizeFailed"));
         }
@@ -532,7 +599,9 @@ export default Vue.extend({
           }
         }
         if (response.status === 200) {
-          this.$toast.success(this.$t("session.notification.messages.issueTrackerSynchronizeSuccess"));
+          this.$toast.success(
+            this.$t("session.notification.messages.issueTrackerSynchronizeSuccess")
+          );
         } else {
           this.$toast.error(this.$t("session.notification.messages.issueTrackerSynchronizeFailed"));
         }
@@ -611,5 +680,82 @@ export default Vue.extend({
   position: relative;
   top: 50%;
   transform: translateY(-50%);
+}
+</style>
+<style scoped>
+.optionButtonCol {
+  margin-top: auto;
+  margin-bottom: auto;
+}
+
+.breakingLine {
+  border-color: var(--text-primary-color);
+}
+
+.copy-popup {
+  text-align: center;
+}
+
+.kick-user {
+  max-height: 500px;
+}
+
+#sessionCloseCol {
+  min-width: 200px;
+}
+
+.headers {
+  display: flex;
+  align-items: center;
+  min-height: 20vh;
+}
+
+.bIcons {
+  height: 40px;
+  width: 40px;
+}
+
+.optionButton {
+  background-color: var(--textAreaColour);
+  color: var(--text-primary-color);
+  border-color: black;
+  border-radius: var(--buttonShape);
+  display: inline-flex;
+  align-items: center;
+}
+
+.optionButton:hover {
+  background-color: var(--textAreaColourHovered);
+  color: var(--text-primary-color);
+}
+
+.optionButton:focus {
+  background-color: var(--textAreaColourHovered) !important;
+  color: var(--text-primary-color) !important;
+}
+
+.refreshButton {
+  border-radius: var(--element-size);
+  color: var(--text-primary-color);
+  background-color: var(--joinButton);
+}
+
+.refreshButton:hover {
+  color: var(--text-primary-color);
+  background-color: var(--joinButtonHovered);
+}
+
+.refreshButton:focus {
+  background-color: var(--joinButtonHovered) !important;
+  color: var(--text-primary-color) !important;
+}
+
+#catGifDiv {
+  text-align: center;
+}
+
+.catGif {
+  width: 240px;
+  height: 180px;
 }
 </style>
